Add animation queue and RedFlash animation
Add a basic circular buffer/queue to store animations. If there is more than one
animation in the queue, each animation will be played for one cycle and removed.
If there is only one animation in the queue, it will be played repeatedly.

Also, add RedFlash animation to be used as a transient animation upon portal
neutralization.